+++ src/shaders/vertex/mod.rs
@@ -1,0 +1,6 @@
+pub mod vs {
+    vulkano_shaders::shader! {
+        ty: "vertex",
+        path: "src/shaders/vertex/base.vert"
+    }
+}